# Markdown Docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bytepath Student Management API
  description: |
    REST API for managing student resources, topics, progress, responses, and reports in the Bytepath learning platform.
    
    This API allows:
    - Authentication and user management
    - Student roster management (CRUD operations, CSV uploads)
    - Topic management and visibility control
    - Progress tracking for students across topics
    - Response tracking for student answers
    - Analytics and reporting for students, topics, and class overview
    
    All endpoints return JSON responses except for the template download endpoint which returns a CSV file.
  version: 1.0.0
  contact:
    name: Bytepath API Support
  license:
    name: MIT
servers:
  - url: http://localhost:5001
    description: Local development server
  - url: https://api.bytepath.example.com
    description: Production server (replace with your actual domain)
...
paths:
...
  /api/students:
    get:
      tags:
        - Students
      summary: Get paginated list of students
      description: |
        Retrieves a paginated list of students. Supports optional search filtering
        on first name, last name, or email (case-insensitive).
      operationId: getStudents
      parameters:
        - name: page
          in: query
          description: Page number for pagination (1-indexed)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: page_size
          in: query
          description: Number of students per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: search
          in: query
          description: Case-insensitive search term to filter by first name, last name, or email
          required: false
          schema:
            type: string
          example: "john"
        - name: include_deleted
          in: query
          description: Include soft-deleted students in results
          required: false
          schema:
            type: boolean
            default: false
        - name: class_name
          in: query
          description: Filter by class name
          required: false
          schema:
            type: string
        - name: sort_by
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            default: "created_at"
        - name: sort_order
          in: query
          description: Sort order (asc or desc)
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: "desc"
      responses:
        '200':
          description: Successful response with paginated student list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Students
      summary: Create a single student manually
      description: |
        Creates a new student record. Email must be unique within a class.

        If a student with the same email was previously removed (soft deleted)
        from the class, their existing record is reactivated and updated with
        any supplied details instead of a new record being created.
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - first_name
                - last_name
              properties:
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                last_name:
                  type: string
                notes:
                  type: string
                class_name:
                  type: string
      responses:
        '200':
          description: Previously removed student reactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterStudent'
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterStudent'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - student is already active in this class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/{id}:
    get:
      tags:
        - Students
      summary: Get single student by ID
      description: Returns a single student record by ID
      operationId: getStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterStudent'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Students
      summary: Partially update a student
      description: |
        Updates student fields. Only provided fields will be updated.
        Used for inline editing in the UI.
      operationId: updateStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                last_name:
                  type: string
                notes:
                  type: string
                class_name:
                  type: string
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterStudent'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Students
      summary: Delete a student
      description: |
        Soft deletes a student by default. Use hard=true query parameter
        for permanent deletion.
      operationId: deleteStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: hard
          in: query
          description: If true, permanently delete instead of soft delete
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Student deleted successfully
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/{id}/restore:
    patch:
      tags:
        - Students
      summary: Restore a soft-deleted student
      description: Restores a previously soft-deleted student
      operationId: restoreStudent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student restored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RosterStudent'
        '400':
          description: Student is not deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/bulk:
    delete:
      tags:
        - Students
      summary: Bulk delete multiple students
      description: Soft deletes multiple students at once
      operationId: bulkDeleteStudents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_ids
              properties:
                student_ids:
                  type: array
                  items:
                    type: integer
                hard:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Bulk delete completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
        '400':
          description: Bad request - student_ids array is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/add:
    post:
      tags:
        - Students
      summary: Add students via CSV upload
      description: |
        Uploads a CSV file to add new students or update existing ones.
        Returns a summary of operations and any validation errors.
      operationId: addStudents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file with student data (columns: first_name, last_name, email)"
      responses:
        '201':
          description: Students added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDropResponse'
        '400':
          description: Bad request - missing file or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/drop:
    post:
      tags:
        - Students
      summary: Drop (soft delete) students via CSV upload
      description: |
        Uploads a CSV file to soft delete students from the roster.
        Returns a summary of operations and any validation errors.
      operationId: dropStudents
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file with student data (columns: first_name, last_name, email)"
      responses:
        '200':
          description: Students dropped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDropResponse'
        '400':
          description: Bad request - missing file or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/upload:
    post:
      tags:
        - Students
      summary: Upload CSV file (legacy endpoint)
      description: |
        Legacy endpoint for backward compatibility. Redirects to /add.
        Uploads a CSV file containing student data.
      operationId: uploadStudentsCsv
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: "CSV file with student data (columns: first_name, last_name, email)"
      responses:
        '200':
          description: Successful upload with processing summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad request - missing file, invalid file type, or empty file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error during file processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/students/upload-history:
    get:
      tags:
        - Students
      summary: Get list of CSV uploads
      description: |
        Returns paginated list of all CSV upload operations with their summaries.
      operationId: getUploadHistory
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Upload history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploads:
                    type: array
                    items:
                      $ref: '#/components/schemas/UploadHistory'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/students/upload-history/{id}:
    get:
      tags:
        - Students
      summary: Get detailed change log for a specific upload
      description: Returns detailed information about a specific CSV upload operation
      operationId: getUploadDetails
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Upload details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadHistory'
        '404':
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/responses/student/{student_id}:
    get:
      tags:
        - Responses
      summary: Get all responses for a student
      description: Returns all response attempts for a given student, newest first
      operationId: getStudentResponses
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Responses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  responses:
                    type: array
                    items:
                      $ref: '#/components/schemas/StudentResponse'

  # Report endpoints
  /api/reports/student/{student_id}:
    get:
      tags:
        - Reports
      summary: Get detailed student performance report
      description: |
        Returns comprehensive performance analytics for a single student,
        including overall stats, per-topic breakdown, performance over time,
        and struggling subtopics.
      operationId: getStudentReport
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Report retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentReport'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    RosterStudent:
      type: object
      description: A student in the roster
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        class_name:
          type: string
          nullable: true
        last_updated_via:
          type: string
          nullable: true
        last_upload_id:
          type: integer
          nullable: true
      required:
        - id
        - email
        - first_name
        - last_name
    StudentListResponse:
      type: object
      description: Paginated response containing a list of students
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RosterStudent'
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 45
        total_pages:
          type: integer
          example: 3
      required:
        - items
        - page
        - page_size
        - total
        - total_pages
    UploadResponse:
      type: object
      description: Response from CSV upload operation (legacy)
      properties:
        summary:
          type: object
          properties:
            inserted:
              type: integer
            updated:
              type: integer
            skipped:
              type: integer
            total_processed:
              type: integer
          required:
            - inserted
            - updated
            - skipped
            - total_processed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
      required:
        - summary
        - errors
    AddDropResponse:
      type: object
      description: Response from add/drop CSV upload operation
      properties:
        upload_id:
          type: integer
        action:
          type: string
          enum: [add, drop]
        summary:
          type: object
          properties:
            inserted:
              type: integer
            updated:
              type: integer
            skipped:
              type: integer
            total_processed:
              type: integer
          required:
            - inserted
            - updated
            - skipped
            - total_processed
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
        upload_history:
          $ref: '#/components/schemas/UploadHistory'
      required:
        - upload_id
        - action
        - summary
        - errors
        - upload_history
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ValidationError:
      type: object
      description: Validation error for a specific CSV row
      properties:
        line:
          type: integer
          description: Line number in the CSV file (1-indexed, excluding header)
          example: 3
        reason:
          type: string
          example: "Missing email"
      required:
        - line
        - reason
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "No file provided"
      required:
        - error